+++ packages/kit/src/core/sync/write_types/index.spec.js
@@ -30,7 +30,7 @@ function run_test(dir) {
 	write_all_types(initial, manifest);
 }
 
-test('Creates correct $types', { timeout: 6000 }, () => {
+test('Creates correct $types', { timeout: 10000 }, () => {
 	// To save us from creating a real SvelteKit project for each of the tests,
 	// we first run the type generation directly for each test case, and then
 	// call `tsc` to check that the generated types are valid.