+++ packages/kit/src/runtime/client/types.d.ts
@@ -125,4 +125,6 @@ export interface HydrateOptions {
 	server_route?: CSRRouteServer;
 	data: Array<ServerDataNode | null>;
 	form: Record<string, any> | null;
+	/** The results of all remote functions executed during SSR so that they can be reused during hydration */
+	remote: Record<string, any>;
 }